Fix the build the right way.
        * platform/StringTruncator.cpp:
        (WebCore::truncateString): Keep all widths as floats.



Canonical link: https://commits.webkit.org/16620@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@19765 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/StringTruncator.cpp
@@ -129,10 +129,10 @@ static String truncateString(const String& string, float maxWidth, const Font& f
         return string;
 
     unsigned keepCountForLargestKnownToFit = 0;
-    unsigned widthForLargestKnownToFit = (unsigned)currentEllipsisWidth;
+    float widthForLargestKnownToFit = currentEllipsisWidth;
     
     unsigned keepCountForSmallestKnownToNotFit = keepCount;
-    unsigned widthForSmallestKnownToNotFit = (unsigned)width;
+    float widthForSmallestKnownToNotFit = width;
     
     if (currentEllipsisWidth >= maxWidth) {
         keepCountForLargestKnownToFit = 1;
@@ -163,10 +163,10 @@ static String truncateString(const String& string, float maxWidth, const Font& f
         width = stringWidth(font, stringBuffer, truncatedLength);
         if (width <= maxWidth) {
             keepCountForLargestKnownToFit = keepCount;
-            widthForLargestKnownToFit = (unsigned)width;
+            widthForLargestKnownToFit = width;
         } else {
             keepCountForSmallestKnownToNotFit = keepCount;
-            widthForSmallestKnownToNotFit = (unsigned)width;
+            widthForSmallestKnownToNotFit = width;
         }
     }
